Separate sla.js into separate file
