+++ package.json
@@ -25,6 +25,7 @@
 		"assets",
 		"disk-icon.icns",
 		"seticon",
+		"sla.js",
 		"base.r"
 	],
 	"keywords": [
